feat: support for netlify added
