+++ package.json
@@ -48,7 +48,7 @@
     "build:server": "vite build --ssr node_modules/rasengan/lib/esm/entries/entry-server.js --outDir dist/server",
     "compile": "tsc -b ./tsconfig.cjs.json ./tsconfig.esm.json ./tsconfig.types.json",
     "build:lib:clean": "rm -rf ./lib",
-    "build:lib:all": "npm-run-all build:lib:clean compile",
+    "build:lib:all": "npm-run-all build:lib:clean compile && node ./src/scripts/copy-extra-files.js",
 
     "deploy": "npm publish --access public",
     "deploy:beta": "npm run deploy --tag beta",
@@ -76,6 +76,7 @@
   },
   "peerDependencies": {
     "@types/node": "^18.0.0 || >=20.0.0",
+    "@netlify/functions": "^2.6.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "sass": "*",
@@ -85,6 +86,9 @@
   },
   "peerDependenciesMeta": {
     "@types/node": {
+      "optional": true
+    },
+    "@netlify/functions": {
       "optional": true
     },
     "sass": {
