+++ src/index.ts
@@ -1,12 +1,13 @@
 // Imports
 import createFetchRequest from "./server/utils/createFetchRequest.js";
+import handleRequest from "./server/utils/handleRequest.js";
 
 // Exports
+// export * from "./decorators/index.js";
 export * from "./core/index.js";
 export * from "./routing/index.js";
-// export * from "./decorators/index.js";
 export * from "./config/index.js";
-export { createFetchRequest }
+export { createFetchRequest, handleRequest }
 
 // Export types
 export type { AppConfig } from "./config/type.js";
