+++ src/scripts/copy-extra-files.js
@@ -1,0 +1,33 @@
+import fs from "node:fs";
+import path from "node:path";
+import { execa } from "execa";
+
+const buildDir = "./lib";
+
+function copyExtraFiles() {
+  fs.readdir(buildDir, function (err, dirs) {
+    if (err) {
+      throw err;
+    }
+    dirs.forEach(function (dir) {
+      if (dir === "esm" || dir === "cjs") {
+        const targetDir = path.join(
+          buildDir,
+          dir
+        );
+        execa(
+          "cp",
+          [
+            `./src/server/functions/netlify/netlify.toml`,
+            `${targetDir}/server/functions/netlify`,
+          ],
+          {
+            stdio: "inherit",
+          }
+        );
+      }
+    });
+  });
+}
+
+copyExtraFiles();
