+++ src/server/utils/handleError.ts
@@ -2,13 +2,6 @@ import path, { join } from "node:path";
 import { Response } from "express";
 
 export function fix404(url: string, res: Response, appPath: string) {
-  // ! Favicon Fix
-  if (url === "/favicon.ico") {
-    return res.sendFile(
-      path.resolve(join(appPath, "dist/client/rasengan.png"))
-    );
-  }
-
   // ! Robots Fix
   if (url === "/robots.txt") {
     return res.send(`
